Reworked code, added new methods
